requirements.txt
cohere
datasets==2.14.6
langchain-text-splitters
langchain
llama-index
pymupdf
pymupdf4llm
unstructured[all-docs]
sentence-transformers
llama-index-embeddings-huggingface
langchain_experimental
pymilvus
pymilvus[model]
pinecone[grpc]
python-dotenv
qdrant-client
fastapi
uvicorn
qdrant-client[fastembed]>=1.8.2
haystack-ai
pinecone-haystack
pathlib
streamlit
pymupdf4llm
pinecone[asyncio]
pydantic-settings